requirements.txt
Gunicorn
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-WTF
psycopg2-binary
python-dotenv
